requirements.txt
--editable src/

# Django 3.2 LTS
django==3.2.*

# Requests for API calls.
requests

# REPL convenience.
ipython

# Rollbar Logging
rollbar==0.1

# Make JSON more readable in admin
django-prettyjson-0.4.*
standardjson-0.3.*